requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.13
aiosignal==1.3.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bcrypt==4.3.0
blinker==1.9.0
cachetools==5.5.2
captcha==0.7.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
dash_cytoscape==1.0.2
dblur==1.0.0
dnspython==2.7.0
efficientnet_pytorch==0.7.1
email_validator==2.2.0
extra-streamlit-components==0.1.71
fastapi==0.115.9
filelock==3.17.0
Flask==3.0.3
frozenlist==1.5.0
fsspec==2025.2.0
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
grpcio==1.70.0
h11==0.14.0
huggingface-hub==0.29.1
idna==3.10
ImageHash==4.3.2
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lightning-utilities==0.12.0
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
narwhals==1.28.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
passlib==1.7.4
pillow==11.1.0
piq==0.8.0
plotly==6.0.0
pretrainedmodels==0.7.4
propcache==0.3.0
protobuf==5.29.3
pyarrow==19.0.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.0.post0
pytz==2025.1
PyWavelets==1.8.0
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
retrying==1.3.4
rich==13.9.4
rpds-py==0.23.1
safetensors==0.5.3
scipy==1.15.2
segmentation_models_pytorch==0.4.0
setuptools==75.8.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.45.3
streamlit==1.42.2
streamlit-authenticator==0.4.2
sympy==1.13.1
tenacity==9.0.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
timm==1.0.15
toml==0.10.2
torch==2.6.0
torchmetrics==1.6.1
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
watchdog==6.0.0
Werkzeug==3.0.6
yarl==1.18.3
zipp==3.21.0
